+++ Nos/Assets/Localization/Localizable.xcstrings
@@ -8377,7 +8377,7 @@
         "en" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "Your @username is your identity in the Nos community. \nChoose a name that reflects you or your organization. Make it memorable and distinct!"
+            "value" : "Your **@username** is your identity in the Nos community. \nChoose a name that reflects you or your organization. Make it memorable and distinct!"
           }
         }
       }
@@ -12734,7 +12734,7 @@
         "en" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "Well done, you've successfully claimed @%1$@! \nYou can share this name with other people in the Nostr and Fediverse communities to make it easy to find you."
+            "value" : "Well done, you've successfully claimed your **@username**! \nYou can share this name with other people in the Nostr and Fediverse communities to make it easy to find you."
           }
         }
       }
