+++ Nos/Views/ProfileEdit/CreateUsernameWizard/CreateUsernameWizard.swift
@@ -1,0 +1,31 @@
+//
+//  CreateUsernameWizard.swift
+//  Nos
+//
+//  Created by Martin Dutra on 31/1/24.
+//
+
+import Dependencies
+import SwiftUI
+
+struct CreateUsernameWizard: ProfileEditWizard {
+
+    @Binding var isPresented: Bool
+
+    @Dependency(\.analytics) private var analytics
+
+    var body: some View {
+        WizardNavigationStack {
+            ClaimYourUniqueIdentitySheet(isPresented: $isPresented)
+        }
+        .onAppear {
+            analytics.showedNIP05Wizard()
+        }
+    }
+}
+
+#Preview {
+    Color.clear.sheet(isPresented: .constant(true)) {
+        CreateUsernameWizard(isPresented: .constant(true))
+    }
+}
