+++ Nos/Views/ProfileEdit/CreateUsernameWizard/PickYourUsernameSheet.swift
@@ -1,0 +1,197 @@
+//
+//  PickYourUsernameSheet.swift
+//  Nos
+//
+//  Created by Martin Dutra on 6/3/24.
+//
+
+import Combine
+import Dependencies
+import Logger
+import SwiftUI
+
+struct PickYourUsernameSheet: ProfileEditSheet {
+
+    @Binding var isPresented: Bool
+    @StateObject private var usernameObserver = UsernameObserver()
+    @State private var verified: Bool?
+    @State private var isVerifying = false
+    @Dependency(\.namesAPI) private var namesAPI
+    @Dependency(\.currentUser) private var currentUser
+
+    var body: some View {
+        SheetVStack {
+            Button {
+                isPresented = false
+            } label: {
+                PlainText(.localizable.cancel)
+                    .font(.clarity(.medium, textStyle: .body))
+                    .foregroundStyle(Color.primaryTxt)
+                    .padding(.vertical, 20)
+                    .padding(.leading, -20)
+            }
+            ScrollView(.vertical, showsIndicators: false) {
+                VStack(alignment: .leading, spacing: 20) {
+                    TitleText(.localizable.pickYourUsernameTitle)
+                    DescriptionText(
+                        AttributedString(
+                            localized: .localizable.pickYourUsernameDescription
+                        )
+                    )
+                    HStack {
+                        UsernameTextField(usernameObserver: usernameObserver)
+                            .onChange(of: usernameObserver.debouncedText) { _, newValue in
+                                Task {
+                                    await verify(newValue)
+                                }
+                            }
+                            .onSubmit {
+                                Task {
+                                    await verify(usernameObserver.text)
+                                }
+                            }
+                        PlainText(".nos.social")
+                            .font(.clarityTitle3)
+                            .foregroundStyle(Color.secondaryTxt)
+                    }
+                    if validationFailed {
+                        usernameAlreadyClaimedText()
+                    } else {
+                        usernameAlreadyClaimedText()
+                            .hidden()
+                    }
+
+                    Spacer(minLength: 0)
+                }
+
+                NavigationLink {
+                    ExcellentChoiceSheet(username: usernameObserver.text, isPresented: $isPresented)
+                } label: {
+                    if isVerifying {
+                        ZStack {
+                            ProgressView()
+                                .frame(height: .zero)
+                                .tint(Color.white)
+                            PlainText(.localizable.next)
+                                .hidden()
+                        }
+                    } else {
+                        PlainText(.localizable.next)
+                    }
+                }
+                .buttonStyle(BigActionButtonStyle())
+                .disabled(verified != true || isVerifying || invalidInput)
+                Spacer(minLength: 40)
+            }
+        }
+    }
+
+    private func usernameAlreadyClaimedText() -> some View {
+        PlainText(.localizable.usernameAlreadyClaimed)
+            .font(.clarity(.medium, textStyle: .subheadline))
+            .foregroundStyle(Color.red)
+            .lineSpacing(3)
+    }
+
+    private var invalidInput: Bool {
+        usernameObserver.text.count < 3
+    }
+
+    private var validationFailed: Bool {
+        verified == false
+    }
+
+    private func verify(_ username: String) async {
+        verified = nil
+
+        guard !username.isEmpty, let keyPair = currentUser.keyPair else {
+            return
+        }
+
+        isVerifying = true
+
+        defer {
+            isVerifying = false
+        }
+
+        do {
+            verified = try await namesAPI.verify(username: username, keyPair: keyPair)
+        } catch {
+            Log.error(error.localizedDescription)
+        }
+    }
+}
+
+fileprivate class UsernameObserver: ObservableObject {
+
+    @Published
+    var debouncedText = ""
+
+    @Published
+    var text = ""
+
+    private var subscriptions = Set<AnyCancellable>()
+
+    init() {
+        $text
+            .removeDuplicates()
+            .filter { $0.count >= 3 }
+            .debounce(for: .milliseconds(200), scheduler: RunLoop.main)
+            .sink { [weak self] value in
+                self?.debouncedText = value
+            }
+            .store(in: &subscriptions)
+    }
+}
+
+fileprivate struct UsernameTextField: View {
+
+    @StateObject var usernameObserver: UsernameObserver
+    @FocusState private var usernameFieldIsFocused: Bool
+
+    var body: some View {
+        SwiftUI.TextField(
+            text: $usernameObserver.text,
+            prompt: PlainText(.localizable.username).foregroundStyle(Color.secondaryTxt)
+        ) {
+            PlainText(.localizable.username)
+                .foregroundStyle(Color.primaryTxt)
+        }
+        .focused($usernameFieldIsFocused)
+        .font(.clarity(.bold, textStyle: .title3))
+        .textInputAutocapitalization(.never)
+        .textCase(.lowercase)
+        .autocorrectionDisabled()
+        .foregroundStyle(Color.primaryTxt)
+        .lineLimit(1)
+        .padding(10)
+        .cornerRadius(10)
+        .overlay {
+            RoundedRectangle(cornerRadius: 10)
+                .stroke(Color.secondaryTxt, lineWidth: 2)
+        }
+        .background {
+            Color.black.opacity(0.1).cornerRadius(10)
+        }
+        .padding(1)
+        .onChange(of: usernameObserver.text) { oldValue, newValue in
+            let characterset = CharacterSet(
+                charactersIn: "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-."
+            )
+            if newValue.rangeOfCharacter(from: characterset.inverted) != nil {
+                usernameObserver.text = oldValue
+            } else if newValue.count > 30 {
+                usernameObserver.text = oldValue
+            } else {
+                usernameObserver.text = newValue.lowercased()
+            }
+        }
+    }
+}
+
+#Preview {
+    Color.clear.sheet(isPresented: .constant(true)) {
+        PickYourUsernameSheet(isPresented: .constant(true))
+            .presentationDetents([.medium])
+    }
+}
