+++ Nos/Views/ProfileEdit/DeleteUsernameWizard/DeleteUsernameWizard.swift
@@ -1,0 +1,34 @@
+//
+//  DeleteUsernameWizard.swift
+//  Nos
+//
+//  Created by Martin Dutra on 6/3/24.
+//
+
+import SwiftUI
+
+struct DeleteUsernameWizard: ProfileEditWizard {
+
+    var author: Author
+    @Binding var isPresented: Bool
+
+    var body: some View {
+        WizardNavigationStack {
+            ConfirmUsernameDeletionSheet(
+                author: author,
+                isPresented: $isPresented
+            )
+        }
+    }
+}
+
+#Preview {
+    var previewData = PreviewData()
+    return Color.clear.sheet(isPresented: .constant(true)) {
+        DeleteUsernameWizard(
+            author: previewData.alice,
+            isPresented: .constant(true)
+        )
+    }
+    .inject(previewData: previewData)
+}
